// src/entities/Articles/model/types/types.ts
export interface ArticleBlockBase {
	id: string
	type: ArticleBlockType
}


export interface ArticleCodeBlock extends ArticleBlockBase {
	type: ArticleBlockType.CODE
	code: string
}


export interface ArticleImageBlock extends ArticleBlockBase{
	type: ArticleBlockType.IMAGE
	src: string
	title: string
}


export interface ArticleTextBlock extends ArticleBlockBase{
	type: ArticleBlockType.TEXT
	paragraphs: string[]
	title?: string
}





export enum ArticleBlockType {
	CODE = 'CODE',
	IMAGE = 'IMAGE',
	TEXT = 'TEXT',
}


export enum ArticleType {
	IT = "IT",
	SCIENCE = "SCIENCE",
	ECONOMICS = "ECONOMICS"
}




export interface Article {
	id: string
	title: string
	subtitle: string
	img: string
	views: number
	createdAt: string
	type: ['IT']
	blocks: ArticleBlock[]
}


export type ArticleBlock = ArticleCodeBlock | ArticleImageBlock | ArticleTextBlock